# Conflicts

+++ nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.app/src/main/java/nts/uk/ctx/at/request/app/command/application/holidayshipment/HolidayShipmentCommand.java
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 public class HolidayShipmentCommand {
 	private String absAppID;
 	private String recAppID;
-	private Long appVersion;
+	private int appVersion;
 	private String memo;
 	private String comboBoxReason;
     private String textAreaReason;
