+++ nts.uk/uk.at/at.ctx/request/nts.uk.ctx.at.request.ws/src/main/java/nts/uk/ctx/at/request/ws/application/common/VersionCheckParam.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class VersionCheckParam {
 	String appID;
-	Long version;
+	int version;
 }
